#911 removed some unused code
