+++ tst/Devices/bluetoothdevicetestdata.h
@@ -6,6 +6,94 @@
 #include <memory>
 #include "devicediscoveryinfo.h"
 
+/**
+ * @brief Device type supported by the application.
+ * NOTE: Not ultimately used for anything currently. This was originally made as part of a plan to split device discovery
+ * into two sections, detection, which returns one of these enum values, and creation, which creates the bluetoothdevice object.
+ * The test data classes still have a function that returns one of these. This has been left here for reuse in a possible future refactoring
+ * of device detection.
+ */
+enum deviceType {
+    None,
+    M3IBike,
+    FakeBike,
+    FakeElliptical,
+    FakeTreadmill,
+    ProformWifiBike,
+    ProformWifiTreadmill,
+    NordicTrackIFitADBTreadmill,
+    NordicTrackIFitADBBike,
+    CSCBike,
+    PowerBike_Stages,
+    PowerTreadmill_StrydrunPowerSensor,
+    DomyosRower,
+    DomyosBike,
+    DomyosElliptical,
+    NautilusElliptical,
+    NautilusBike,
+    ProformElliptical,
+    NordicTrackElliptical,
+    ProformEllipticalTrainer,
+    ProformRower,
+    BHFitnessElliptical,
+    SoleElliptical,
+    DomyosTreadmill,
+    KingsmithR2Treadmill,
+    KingsmithR1ProTreadmill,
+    Shuaa5Treadmill,
+    TrueTreadmill,
+    SoleF80Treadmill,
+    HorizonTreadmill,
+    TechnoGymMyRunTreadmill,
+    TechnoGymMyRunTreadmillRfComm,
+    TacxNeo2,
+    NPECableBike,
+    FTMSBike,
+    WahooKickrSnapBike,
+    HorizonGr7Bike,
+    StagesBike,
+    SmartRowRower,
+    Concept2SkiErg,
+    FTMSRower,
+    EchelonStride,
+    OctaneElliptical,
+    OctaneTreadmill,
+    EchelonRower,
+    EchelonConnectSport,
+    SchwinnIC4Bike,
+    SportsTechBike,
+    SportsPlusBike,
+    YesoulBike,
+    ProformBike,
+    ProformTreadmill,
+    ESLinkerTreadmill,
+    PafersTreadmill,
+    BowflexT216Treadmill,
+    NautilusTreadmill,
+    FlywheelBike,
+    MCFBike,
+    ToorxTreadmill,
+    IConceptBike,
+    SpiritTreadmill,
+    ActivioTreadmill,
+    TrxAppGateUSBTreadmill,
+    TrxAppGateUSBBike,
+    UltraSportBike,
+    KeepBike,
+    SoleBike,
+    SkandikawiriBike,
+    RenphoBike,
+    PafersBike,
+    SnodeBike,
+    FitPlusBike,
+    FitshowTreadmill,
+    InspireBike,
+    ChronoBike,
+    MepanelBike,
+    LifeFitnessTreadmill,
+    CompuTrainerBike
+};
+
 class BluetoothDeviceTestData;
 typedef std::shared_ptr<BluetoothDeviceTestData> BluetoothDeviceTestData_ptr;
 typedef std::vector<BluetoothDeviceTestData_ptr> BluetoothDeviceTestDataVector;
